requirements.txt
Django>=5.0,<5.2
djangorestframework>=3.15,<3.16
django-filter>=24.2
psycopg2-binary>=2.9
django-environ>=0.11
Pillow>=10.3
django-ratelimit>=4.1
django-storages>=1.14
boto3>=1.34
whitenoise>=6.6
bleach>=6.1
gunicorn>=22.0
